Create combo dataset
In this commit, we create a new dataset to train and test a
metonymy-aware entity linking system.

We choose to use AIDA and WiMCor datasets for now. However it is
better to stick to WiMCor itself. Since it takes time to learn
entity embeddings, we decided to use AIDA for now because AIDA
entities already exist in the entity universe.

We include the feature that specifies whether the entity is
literal (LIT) or metonymic (MET).