Improve AIDA and WiMCor preprocessing
The preprocessing routines directly directed the output to file.
In a previous commit, I have modified these routines to save the
output to a list as well to create the combo dataset.

In this commit, I have cleaned up the routines so that output is
NOT immediately written to file but saved and written to file at
a later stage.